+++ frontend/packages/kubevirt-plugin/src/plugin.tsx
@@ -131,7 +131,6 @@ const plugin: Plugin<ConsumedExtensions> = [
       exact: true,
       path: ['/k8s/ns/:ns/vmtemplates/~new'],
       loader: () =>
-        /* eslint-disable-next-line import/no-cycle */
         import(
           './components/vm-templates/vm-template-create-yaml' /* webpackChunkName: "kubevirt-vmtemplates-create-yaml" */
         ).then((m) => m.CreateVMTemplateYAML),
